refactor(subagents-v3): rename widget module
