+++ packages/pi-subagents-v3/src/widget.ts
