+++ packages/pi-subagents-v3/test/subagents.test.ts
@@ -6,9 +6,9 @@ import { afterEach, beforeEach, test, vi } from "vitest";
 import { createMockContext, createMockPi } from "../../../test/support.js";
 import { createBrokerClient } from "../src/child-communication-bridge.js";
 import { MessageBroker } from "../src/message-broker.js";
-import { SUBAGENT_WIDGET_KEY } from "../src/subagent-widget.js";
 import subagentsV3, { type SubagentsV3Dependencies } from "../src/subagents.js";
 import type { ChildRequest, ChildResult } from "../src/types.js";
+import { SUBAGENT_WIDGET_KEY } from "../src/widget.js";
 
 interface RegisteredTool {
 	name: string;
